requirements.txt
streamlit
mem0ai
python-dotenv
requests
